% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdrift.R
\docType{package}
\name{bdrift-package}
\alias{bdrift}
\alias{bdrift-package}
\title{Analyzing Beta Drift}
\description{
bdrift provides tools for analyzing and
visualizing beta drift in multi-factor models.
}

