% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.BDA}
\alias{print.BDA}
\title{Print beta-drift anaylses}
\usage{
\method{print}{BDA}(x, ...)
}
\arguments{
\item{x}{an object of class \code{BDA}.}

\item{...}{additional parameters.}
}
\description{
\code{print.BDA} prints beta-drift anaylses.
}
\details{
This function prints simplified summary statistics of analyses
created by the \code{\link{BDA}} function.
}
\examples{
results <- BDA(data = FFfactors, spec = (VOO~SP500),
horizon = 250, doplot = FALSE)
print(results)
}
\author{
Markus Peter Auer <mp.auer@meanerreversion.com>
}

