% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.BDA}
\alias{summary.BDA}
\title{Summarize beta-drift analyses}
\usage{
\method{summary}{BDA}(object, ...)
}
\arguments{
\item{object}{an object of class \code{BDA}.}

\item{...}{additional parameters.}
}
\description{
\code{summary.BDA} summarizes the results of beta-drift analyses.
}
\details{
This function prints a detailed summary of the analyses produced by
the \code{\link{BDA}} function.

The summary is threefold.fj
}
\examples{
results <- BDA(data = FFfactors, spec = (MDLRX~Mkt.RF + SMB + HML + RMW + CMA),
horizon = 250, doplot = FALSE)
summary(results)
}
\author{
Markus Peter Auer <mp.auer@meanerreversion.com>
}

